+++ elf_examples/add_and_square.c
@@ -1,0 +1,4 @@
+int add_and_square(int a, int b) {
+    int sum = a + b;
+    return sum * sum;
+}